% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDCVBlockBootMulti.R
\name{cdcv.multi.test}
\alias{cdcv.multi.test}
\title{Full test of independence of multivariate time series X and Y.
Uses cross-distance covariance and circular or stationary bootstrap.
For univariate data, use cdcv.univ.test.}
\usage{
cdcv.multi.test(X, Y, M = NA, num_boot = 100, alpha = 0.05,
  unbiased = FALSE, type = "stationary", block_size = NA)
}
\arguments{
\item{X}{Time series as n x d_X matrix.}

\item{Y}{Time series as n x d_Y matrix.}

\item{M}{Maximum lag to consider for cross-distance covariance.}

\item{num_boot}{Number of bootstrapped samples.}

\item{alpha}{Significance level of test.}

\item{unbiased}{Whether to use the biased or unbiased estimate of dCov.}

\item{type}{"circular" or "stationary" boostrap.}

\item{block_size}{Block size or mean block size.}
}
\value{
A list containing the following:
\item{\code{pCDCV}}{P-value of CDCV}
\item{\code{statCDCV}}{test statistic value}
\item{\code{unbiased}}{Whether to use the unbiased or biased estimate of CDCV}
\item{\code{M}}{the max lead/lag considered}
}
\description{
Full test of independence of multivariate time series X and Y.
Uses cross-distance covariance and circular or stationary bootstrap.
For univariate data, use cdcv.univ.test.
}
