% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkTest.R
\name{DiffusionTest}
\alias{DiffusionTest}
\title{Distance-based graph test using diffusion map embedding}
\usage{
DiffusionTest(G, X, option = "mgc", t.max = 5, n.perm = 100,
  default.q = 10)
}
\arguments{
\item{G}{the input igraph object.}

\item{X}{the input nodal attributes.}

\item{option}{is a string that specifies which global correlation to build up-on. Defaults to \code{'mgc'}.
\describe{
   \item{\code{'mgc'}}{uses the MGC global correlation.}
   \item{\code{'dcor'}}{uses the dcor global correlation.}
   \item{\code{'hhg'}}{uses Heller-Heller-Gorfine (HHG) tests.}
}}

\item{t.max}{the maximum diffusion time. It starts from 0, and default is \code{t.max} = 5.}

\item{n.perm}{the number of permutation samples.  Defaults to \code{n.perm} = 100.}

\item{default.q}{a dimension of graph embeddings. Defaults to \code{default.q} = 10.}
}
\value{
A list containing the following:
\item{\code{stats}}{a set of multiscale statistics across diffusion times.}
\item{\code{stats.null}}{a \code{[n.perm x (t.max + 1)]} matrix of multiscale statistics under the null.}
\item{\code{dimension}}{a dimension of diffusion map embedding.}
\item{\code{localmap}}{a list of local correlation map (in case of \code{option} = 'mgc').}
\item{\code{localscale}}{a list of optimal local scale (row, column) in each local correlation map (in case of \code{option} = 'mgc').}
}
\description{
An auxiliary function to print out multiscale statistics and those under the permutation sample
along with local correlations map across diffusion time.
}
\author{
Youjin Lee
}
