% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{discr.discr}
\alias{discr.discr}
\title{Discriminability
Mean Normalized Rank}
\usage{
discr.discr(D, ids, remove_outliers = TRUE, thresh = 0, output = FALSE)
}
\arguments{
\item{D}{[n, n] a distance matrix for n subjects.}

\item{ids}{[n]: a vector containing the subject ids for each subject.}

\item{remove_outliers=TRUE}{boolean indicating whether to ignore subjects with rdf below a certain cutoff.}

\item{thresh=0}{[1]: the threshold below which to ignore subjects.}

\item{output=FALSE}{a boolean indicating whether to ignore output.}
}
\value{
discr [1]: the discriminability statistic.
}
\description{
A function for computing the discriminability from a distance matrix and a set of associated labels.
}
\seealso{
{discr.distance}
}
\author{
Eric Bridgeford and Gregory Kiar
}
