% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{discr.distance}
\alias{discr.distance}
\title{Distance}
\usage{
discr.distance(X, method = "2")
}
\arguments{
\item{X}{[n, d] a data matrix for n observations of d variables.}

\item{method='2'}{the method for computing distances.
\describe{
   \item{'2'}{use the 2 norm between pairs of observations}
}}
}
\value{
dist [n, n]: a matrix indicating the pairwise distances between all observations passed in.
}
\description{
A function that returns a distance matrix given a collection of observations.
}
\author{
Eric Bridgeford
}
