% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrKDE.R
\name{discr.kde_subject}
\alias{discr.kde_subject}
\title{KDE of Subject Relationships}
\usage{
discr.kde_subject(D, ids)
}
\arguments{
\item{D}{[n, n]: a distance matrix corresponding to the distances between n scans.}

\item{ids}{[n]: a vector where each entry is the subject id of each of the n scans.}
}
\value{
f_intra [x]: a density estimate of the intra subject relationship.

f_inter [x]: a density estimate of the inter subject relationship.

H [1]: the hellinger distance between the intra and inter subject relationships.
}
\description{
A function that  returns density estimates of intra and inter subject distances.
}
\author{
Eric Bridgeford
}
