% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrStat.R
\name{discr.mnr}
\alias{discr.mnr}
\title{Mean Normalized Rank}
\usage{
discr.mnr(rdf, remove_outliers = TRUE, thresh = 0, output = FALSE)
}
\arguments{
\item{rdf}{[n]: the reliability density function.}

\item{remove_outliers=TRUE}{boolean indicating whether to ignore subjects with rdf below a certain cutoff.}

\item{thresh=0}{[1]: the threshold below which to ignore subjects.}

\item{output=FALSE}{a boolean indicating whether to ignore output.}
}
\value{
discr [1]: the discriminability statistic.
}
\description{
A function for computing the mnr from an rdf.
}
\author{
Eric Bridgeford and Gregory Kiar
}
