% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_data.R
\name{discr.rank_data}
\alias{discr.rank_data}
\title{Rank Data}
\usage{
discr.rank_data(X, normalize = FALSE)
}
\arguments{
\item{X}{[n, d] the data to rank, with n observations of d variables.}

\item{normalize=FALSE}{: a boolean indicating whether to normalize so the rank for a particular observation has values betweeen 0 and 1.}
}
\value{
rX [n, d] the ranked data, where the ith row contains the ranks of the ith set of observations.
}
\description{
A function that ranks all the variables for a particular observation in a dataset.
}
\author{
Eric Bridgeford
}
