% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrStat.R
\name{discr.rdf}
\alias{discr.rdf}
\title{Reliability Density Function}
\usage{
discr.rdf(D, ids)
}
\arguments{
\item{D}{[n, n] a distance matrix for n subjects.}

\item{ids}{[n]: a vector containing the subject ids for each subject.}
}
\value{
rdf [n]: the reliability per subject.
}
\description{
A function for computing the reliability density function of a dataset.
}
\author{
Shangsi Wang, Eric Bridgeford and Gregory Kiar
}
