% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo.R
\name{discr.run_demo}
\alias{discr.run_demo}
\title{A simple demo function to illustrate Discriminability on 2 realizations of 2 random variables.}
\usage{
discr.run_demo()
}
\description{
A simple demo function to illustrate Discriminability on 2 realizations of 2 random variables.
}
