% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{discr.sims.beta}
\alias{discr.sims.beta}
\title{Discriminability Beta Simulation}
\usage{
discr.sims.beta(n, d, class.equal = TRUE, ind = FALSE)
}
\arguments{
\item{n}{the number of samples.}

\item{d}{the number of dimensions.}

\item{class.equal}{whether the number of samples/class should be equal, with each
class having a prior of 1/K, or inequal, in which each class obtains a prior
of k/sum(K) for k=1:K. Defaults to \code{TRUE}.}

\item{ind}{whether to sample x and y independently. Defaults to \code{FALSE}.}
}
\description{
Signal Dimension is beta-distributed; remainder are Uniform 0, 1
}
\author{
Eric Bridgeford
}
