% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{discr.sims.spread}
\alias{discr.sims.spread}
\title{Discriminability Spread Simulation}
\usage{
discr.sims.spread(n, d, K, sig.scale = 1, rotate = FALSE,
  class.equal = TRUE, ind = FALSE)
}
\arguments{
\item{n}{the number of samples.}

\item{d}{the number of dimensions.}

\item{K}{the number of classes in the dataset.}

\item{sig.scale}{the scaling for the class-variance. Defaults to \code{1}.}

\item{class.equal}{whether the number of samples/class should be equal, with each
class having a prior of 1/K, or inequal, in which each class obtains a prior
of k/sum(K) for k=1:K. Defaults to \code{TRUE}.}

\item{ind}{whether to sample x and y independently. Defaults to \code{FALSE}.}
}
\description{
A function to simulate data with the same mean that spreads as class id increases.
}
\examples{
library(mgc)
sim <- discr.sims.spread(100, 3, 2)
}
\author{
Eric Bridgeford
}
