% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrPermutationTests.R
\name{discr.test.two_sample}
\alias{discr.test.two_sample}
\title{Discriminability Two Sample Permutation Test}
\usage{
discr.test.two_sample(X1, X2, Y, dist.xfm = discr.distance,
  dist.params = list(method = "euclidian"), dist.return = NULL,
  remove.isolates = TRUE, nperm = 100, no_cores = 1,
  alt = "greater")
}
\arguments{
\item{X1}{is interpreted as a \code{[n x d]} data matrix with \code{n} samples in \code{d} dimensions. Should NOT be a distance matrix.}

\item{X2}{is interpreted as a \code{[n x d]} data matrix with \code{n} samples in \code{d} dimensions. Should NOT be a distance matrix.}

\item{Y}{\code{[n]} a vector containing the sample ids for our \code{n} samples. Should be matched such that \code{Y[i]} is the corresponding label for \code{X1[i,]} and \code{X2[i,]}.}

\item{dist.xfm}{if \code{is.dist == FALSE}, a distance function to transform \code{X}. If a distance function is passed,
it should accept an \code{[n x d]} matrix of \code{n} samples in \code{d} dimensions and return a \code{[n x n]} distance matrix
as the \code{$D} return argument. See \link[mgc]{discr.distance} for details.}

\item{dist.params}{a list of trailing arguments to pass to the distance function specified in \code{dist.xfm}.
Defaults to \code{list(method='euclidean')}.}

\item{dist.return}{the return argument for the specified \code{dist.xfm} containing the distance matrix. Defaults to \code{FALSE}.
\describe{
    \item{\code{is.null(dist.return)}}{use the return argument directly from \code{dist.xfm} as the distance matrix. Should be a \code{[n x n]} matrix.}
    \item{\code{is.character(dist.return) | is.integer(dist.return)}}{use \code{dist.xfm[[dist.return]]} as the distance matrix. Should be a \code{[n x n]} matrix.}
}}

\item{remove.isolates}{remove isolated samples from the dataset. Isolated samples are samples with only
one instance of their class appearing in the \code{Y} vector. Defaults to \code{TRUE}.}

\item{nperm}{the number of permutations for permutation test. Defualts to \code{100}.}

\item{no_cores}{the number of cores to use for the permutations. Defaults to \code{1}.}

\item{alt}{the alternative hypothesis. Can be that first dataset is more discriminable (\code{alt = 'greater'}), less discriminable (\code{alt = 'less'}),
or just non-equal (\code{alt = 'neq'}). Defaults to \code{"greater"}.}
}
\value{
A list containing the following:
\item{\code{stat}}{the observed test statistic.}
\item{\code{p.value}}{The p-value associated with the test.}
}
\description{
A function that takes two distance matrices and produces a p-value associated with whether or not the distance matrices differ significantly.
}
\author{
Eric Bridgeford
}
