% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCDistTransform.R
\name{DistRanks}
\alias{DistRanks}
\title{An auxiliary function that sorts the entries within each column by ascending order:
For ties, the minimum ranking is used,
e.g. if there are repeating distance entries, the order is like 1,2,3,3,4,..,n-1.}
\usage{
DistRanks(dis)
}
\arguments{
\item{dis}{is a symmetric distance matrix.}
}
\value{
\code{disRank} is the column rank matrices of \code{X}.
}
\description{
An auxiliary function that sorts the entries within each column by ascending order:
For ties, the minimum ranking is used,
e.g. if there are repeating distance entries, the order is like 1,2,3,3,4,..,n-1.
}
\keyword{internal}
