% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkTest.R
\name{dmap.q}
\alias{dmap.q}
\title{Diffusion maps}
\usage{
dmap.q(L, times = 1, q = 10)
}
\arguments{
\item{L}{[n x n] Laplacian matrix.}

\item{times}{a vector of diffusion times. Defaults to 1.}

\item{q}{dimension of diffusion map embeddings.}
}
\value{
a list of diffusion map embeddings with length [q].
}
\description{
This function provides a set of diffusion maps.
}
\examples{

library(igraphdata)
library(mgc)
data(karate)
G <- karate
A <- as.matrix(get.adjacency(G))
D <- diag(pmin( (rowSums(A))^(-1/2) , 1))
P <- D \%*\% A \%*\% D
diffusion.map <- dmap.q(P, c(0:5), q = 5)

}
\author{
Youjin Lee
}
