% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gen.coefs}
\alias{gen.coefs}
\title{A helper function to generate a d-dimensional linear transformation matrix.}
\usage{
gen.coefs(d)
}
\description{
A helper function to generate a d-dimensional linear transformation matrix.
}
