% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gen.sample.labels}
\alias{gen.sample.labels}
\title{A helper function for simulating sample labels}
\usage{
gen.sample.labels(K, class.equal = TRUE)
}
\arguments{
\item{K}{the number of classes}

\item{class.equal}{whether the number of samples/class should be equal, with each
class having a prior of 1/K, or inequal, in which each class obtains a prior
of k/sum(K) for k=1:K. Defaults to \code{TRUE}.}
}
\description{
A helper function for simulating sample labels
}
\keyword{internal}
