% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gen.x}
\alias{gen.x}
\title{A helper function to generate n samples of a d-dimensional uniform vector.}
\usage{
gen.x(n, d, a = -1, b = 1)
}
\description{
A helper function to generate n samples of a d-dimensional uniform vector.
}
