% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gen.x.unif}
\alias{gen.x.unif}
\title{A helper function to generate n samples of a d-dimensional uniform vector.}
\usage{
gen.x.unif(n, d, a = -1, b = 1)
}
\arguments{
\item{n}{the number of samples.}

\item{d}{the number of dimensions.}

\item{a}{the lower limit.}

\item{b}{the upper limit.}

\item{x}{\code{[n, d]} the simulated data matrix.}
}
\description{
A helper function to generate n samples of a d-dimensional uniform vector.
}
\author{
Eric Bridgeford
}
\keyword{internal}
