% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCLocalCorr.R
\name{LocalCov}
\alias{LocalCov}
\title{An auxiliary function that computes all local correlations simultaneously in O(n^2).}
\usage{
LocalCov(A, B, RX, RY)
}
\arguments{
\item{A}{is a properly transformed distance matrix}

\item{B}{is the second distance matrix properly transformed}

\item{RX}{is the column-ranking matrix of A}

\item{RY}{is the column-ranking matrix of B.}
}
\value{
covXY is all local covariances computed iteratively.
}
\description{
An auxiliary function that computes all local correlations simultaneously in O(n^2).
}
\keyword{internal}
