% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCDistTransform.R
\name{mgc.dist}
\alias{mgc.dist}
\title{Transform the distance matrices, with column-wise ranking if needed.}
\usage{
mgc.dist(X, Y, option = "mgc", optionRk = TRUE)
}
\arguments{
\item{X}{[nxn] is a distance matrix;}

\item{Y}{[nxn] is a second distance matrix;}

\item{option='mgc'}{is a string that specifies which global correlation to build up-on.
\describe{
   \item{'mgc'}{use the MGC global correlation.}
   \item{'dcor'}{use the dcor global correlation.}
   \item{'mantel'}{use the mantel global correlation.}
   \item{'rank'}{use the rank global correlation.}
}}

\item{optionRk=TRUE}{is a string that specifies whether ranking within column is computed or not. If option='rank', ranking will be performed regardless of the value specified by optionRk.}
}
\value{
A [nxn] the centered distance matrix for X.

B [nxn] the centered distance matrix for Y.

RX [nxn] the column-rank matrices of X.

RY [nxn] the column-rank matrices of Y.
}
\description{
Transform the distance matrices, with column-wise ranking if needed.
}
