% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCDistTransform.R
\name{mgc.distance}
\alias{mgc.distance}
\title{Distance}
\usage{
mgc.distance(X, method = "euclidean")
}
\arguments{
\item{X}{\code{[n x d]} a data matrix for \code{d} samples of \code{d} variables.}

\item{method}{the method for computing distances. Defaults to \code{'euclidean'}. See \link[stats]{dist} for details. Also
includes a "ohe" option, which one-hot-encodes the matrix when computing distances.}
}
\value{
a \code{[n x n]} distance matrix indicating the pairwise distances between all samples passed in.
}
\description{
A function that returns a distance matrix given a collection of observations.
}
\author{
Eric Bridgeford
}
