% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCKSample.R
\name{mgc.ksample}
\alias{mgc.ksample}
\title{This function tests the equivalence of K distributions using distance-based test including MGC.
Input data from K populations are first transformed into pooled data Z and label vector Y.
then implements test statistics upon distance matrices from each data X and Y.}
\usage{
mgc.ksample(dat, dist = "Euclidean")
}
\arguments{
\item{dat}{is interpreted as:
\describe{
   \item{list of length [k]}{dat is a k-length list of the same size of vectors or matrices}
}}

\item{dist='Euclidean'}{is a string that specifies distance matrix to apply.}
}
\value{
result : list of results of mgc.stat testing equivalence of K-sample test.
}
\description{
This function tests the equivalence of K distributions using distance-based test including MGC.
Input data from K populations are first transformed into pooled data Z and label vector Y.
then implements test statistics upon distance matrices from each data X and Y.
}
\author{
Youjin Lee
}
