% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCLocalCorr.R
\name{mgc.localcorr}
\alias{mgc.localcorr}
\title{MGC Local Correlations}
\usage{
mgc.localcorr(X, Y, option = "mgc")
}
\arguments{
\item{X}{is interpreted as:
\describe{
   \item{a [n x n] distance matrix}{X is a square matrix with zeros on diagonal}
   \item{a [n x d] data matrix}{Otherwise}
}}

\item{Y}{is interpreted as:
\describe{
   \item{a [n x n] distance matrix}{Y is a square matrix with zeros on diagonal}
   \item{a [n x d] data matrix}{Otherwise}
}}

\item{option='mgc'}{is a string that specifies which global correlation to build up-on.
\describe{
   \item{'mgc'}{use the MGC global correlation.}
   \item{'dcor'}{use the dcor global correlation.}
   \item{'mantel'}{use the mantel global correlation.}
   \item{'rank'}{use the rank global correlation.}
}}
}
\value{
A list contains the following output:

corr consists of all local correlations within [-1,1] by double matrix index;

varX contains all local variances for X; varY contains all local covariances for Y.
}
\description{
Compute all local correlation coefficients in O(n^2 log n)
}
\author{
C. Shen
}
