% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{mgc.plot.plot_matrix}
\alias{mgc.plot.plot_matrix}
\title{Matrix plot}
\usage{
mgc.plot.plot_matrix(mtx, title = "", xlabel = "", ylabel = "",
  legend.name = "metric", legend.show = TRUE, font.size = 12,
  limits = NaN)
}
\arguments{
\item{mtx}{[nroi, nroi] or [nt, nroi]: the input. Can either be a square matrix of featuresxfeatures, or a data matrix of observationsxfeatures.}

\item{title="":}{the title for the square plot.}

\item{xlabel="":}{the x label for the square plot.}

\item{ylabel="":}{the y label for the square plot.}

\item{legend.name="":}{the legend title for the square plot.}

\item{legend.show=TRUE:}{whether to show the legend on the plot.}

\item{font.size=12:}{the default font size for the plot text. Axis/legend text is font.size - 2.}
}
\description{
A function that plots a matrix.
}
\author{
Eric Bridgeford
}
