% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{mgc.plot.plot_mgc}
\alias{mgc.plot.plot_mgc}
\title{MGC plot}
\usage{
mgc.plot.plot_mgc(mtx, title = "MGC Power Map",
  xlabel = TeX("$k$, $X$ Neighbors"), ylabel = TeX("$l$, $Y$ Neighbors"),
  legend.name = "statMGC", legend.show = TRUE, font.size = 12,
  limits = NaN)
}
\arguments{
\item{mtx}{[n, n]: the input. Should be an output from mgc.sample(...)$localCorr}

\item{title="MGC}{Power Map": the title for the square plot.}

\item{xlabel=TeX("$k$, }{$X$ neighbors"): the x label for the square plot.}

\item{ylabel=TeX("$l$, }{$Y$ neighbors"): the y label for the square plot.}

\item{legend.name="statMGC":}{the legend title for the square plot.}

\item{legend.show=TRUE:}{whether to show the legend on the plot.}

\item{font.size=12:}{the default font size for the plot text. Axis/legend text is font.size - 2.}
}
\description{
A function that plots a MGC local correlation/power map.
}
\seealso{
{mgc.sample}
}
\author{
Eric Bridgeford
}
