% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo.R
\name{mgc.run_demo}
\alias{mgc.run_demo}
\title{A simple demo function to illustrate MGC on linear and quadratic relationships at size 20.}
\usage{
mgc.run_demo()
}
\description{
A simple demo function to illustrate MGC on linear and quadratic relationships at size 20.
}
