% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCSampleStat.R
\name{mgc.sample}
\alias{mgc.sample}
\title{MGC Sample}
\usage{
mgc.sample(A, B, option = "mgc")
}
\arguments{
\item{A}{is interpreted as:
\describe{
   \item{a \code{[nxn]} distance matrix}{A is a square matrix with zeros on diagonal}
   \item{a \code{[nxd]} data matrix}{Otherwise}
}}

\item{B}{is interpreted as:
\describe{
   \item{a \code{[nxn]} distance matrix}{A is a square matrix with zeros on diagonal}
   \item{a \code{[nxd]} data matrix}{Otherwise}
}}

\item{mgc}{is a string that specifies which global correlation to build up-on. Defaults to \code{'mgc'}.
\describe{
   \item{\code{'mgc'}}{use the MGC global correlation.}
   \item{\code{'dcor'}}{use the dcor global correlation.}
   \item{\code{'mantel'}}{use the mantel global correlation.}
   \item{\code{'rank'}}{use the rank global correlation.}
}}
}
\value{
A list containing the following:
\item{\code{statMGC}}{is the sample MGC statistic within \code{[-1,1]}}
\item{\code{localCorr}}{consists of all local correlations by double matrix index}
\item{\code{optimalScale}}{the estimated optimal scale in matrix single index.}
}
\description{
The main function that computes the MGC measure between two datasets:
It first computes all local correlations,
then use the maximal statistic among all local correlations based on thresholding.
}
\author{
C. Shen
}
