% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{mgc.sims.joint}
\alias{mgc.sims.joint}
\title{Joint Normal}
\usage{
mgc.sims.joint(n, d, eps = 0.5)
}
\arguments{
\item{n}{the number of samples for the simulation.}

\item{d}{the number of dimensions for the simulation setting.}

\item{eps=0.5}{the noise level for the simulation.}
}
\value{
X [n, d] the data matrix.

Y [n] the response array.
}
\description{
A function for Generating a Joint Normal Simulation.
}
\author{
Eric Bridgeford
}
