% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{mgc.sims.linear}
\alias{mgc.sims.linear}
\title{Linear Simulation}
\usage{
mgc.sims.linear(n, d, eps = 1, ind = FALSE, a = -1, b = -1)
}
\arguments{
\item{n}{the number of samples for the simulation.}

\item{d}{the number of dimensions for the simulation setting.}

\item{eps}{the noise level for the simulation. Defaults to \code{1}.}

\item{ind}{whether to sample x and y independently. Defaults to \code{FALSE}.}

\item{a}{the lower limit for the range of the data matrix. Defaults to \code{-1}.}

\item{b}{the upper limit for the range  of the data matrix. Defaults to \code{1}.}
}
\value{
a list containing the following:
\itemize{
\item{\code{X}}{\code{[n, d]} the data matrix with \code{n} samples in \code{d} dimensions.}
\item{\code{Y}}{\code{[n]} the response array.}
}
}
\description{
A function for Generating a Linear Simulation.
}
\section{Details}{

\eqn{Linear(X, Y) \in  \mathcal{R}^d \cross \mathcal{R}}{Linear(X, Y)}, where:
\deqn{X \sim \mathcal{U}(a, b)^d}{X ~ U(a, b)^d}
\deqn{Y = w^TX + \kappa \epsilon}{Y = w^T X + \kappa \epsilon}
and
\deqn{\kappa = \begin{cases}1 & d = 1 \\ 0 & d > 1\end{cases}}{\kappa = 1 if d = 1, and 0 otherwise}
}

\author{
Eric Bridgeford
}
