% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{mgc.sims.wshape}
\alias{mgc.sims.wshape}
\title{W Shaped Function}
\usage{
mgc.sims.wshape(n, d, eps = 0.5, ind = FALSE, a = -1, b = -1)
}
\arguments{
\item{n}{the number of samples for the simulation.}

\item{d}{the number of dimensions for the simulation setting.}

\item{eps=0.5}{the noise level for the simulation.}

\item{ind=FALSE}{whether to sample x and y independently.}

\item{a=-1}{the lower limit for the data matrix.}

\item{b=1}{the upper limit for the data matrix.}
}
\value{
a list containing the following:
\item{\code{X}}{\code{[n, d]} the data matrix with \code{n} samples in \code{d} dimensions.}
\item{\code{Y}}{\code{[n]} the response array.}
}
\description{
A function for Generating a W Simulation.
}
\author{
Eric Bridgeford
}
