% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCDistTransform.R
\name{MGCDistTransform}
\alias{MGCDistTransform}
\title{Transform the distance matrices, with column-wise ranking if needed.}
\usage{
MGCDistTransform(X, Y, option, optionRk)
}
\arguments{
\item{X}{is a distance matrix;}

\item{Y}{is a second distance matrix;}

\item{option}{is a string that specifies which global correlation to build up-on, including 'mgc','dcor','mantel', and 'rank';}

\item{optionRk}{is a string that specifies whether ranking within column is computed or not.}
}
\value{
A list contains the following output:

A and B are the centered distance matrices;

RX and RY are the column rank matrices of X and Y respectively.
}
\description{
Transform the distance matrices, with column-wise ranking if needed.
}
