% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCLocalCorr.R
\name{MGCLocalCorr}
\alias{MGCLocalCorr}
\title{Compute all local correlation coefficients in O(n^2 log n)}
\usage{
MGCLocalCorr(X, Y, option)
}
\arguments{
\item{X}{is a distance matrix or a n*d data matrix; if it is not a square matrix with zeros on diagonal, it is treated as n*d data;}

\item{Y}{is a second distance matrix or a n*d data matrix, with the same distance matrix check as X;}

\item{option}{is a string that specifies which global correlation to build up-on, including 'mgc','dcor','mantel', and 'rank'.}
}
\value{
A list contains the following output:

corr consists of all local correlations within [-1,1] by double matrix index;

varX contains all local variances for X; varY contains all local covariances for Y.
}
\description{
Compute all local correlation coefficients in O(n^2 log n)
}
