% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCPermutationTest.R
\name{MGCPermutationTest}
\alias{MGCPermutationTest}
\title{The main function that tests independent between two data sets by MGC and permutation test.}
\usage{
MGCPermutationTest(X, Y, rep, option)
}
\arguments{
\item{X}{is a distance matrix or a n*d data matrix; if it is not a square matrix with zeros on diagonal, it is treated as n*d data;}

\item{Y}{is a second distance matrix or a n*d data matrix, with the same distance matrix check as X;}

\item{rep}{specifies the number of replicates to use for the permutation test;}

\item{option}{is a string that specifies which global correlation to build up-on, including 'mgc','dcor','mantel', and 'rank'.}
}
\value{
A list contains the following output:

P-value and test statistic of MGC;

All local p-values by double matrix index, all local correlations by double matrix index, and the estimated optimal scales as matrix single indices.

Note that one should avoid report positive discovery via minimizing individual p-values of local correlations,
unless corrected for multiple testing problem.
}
\description{
The main function that tests independent between two data sets by MGC and permutation test.
}
