% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCSampleStat.R
\name{MGCSampleStat}
\alias{MGCSampleStat}
\title{The main function that computes the MGC measure between two datasets:
It first computes all local correlations,
then thresholding and use the smoothed maximal of the local correlations.}
\usage{
MGCSampleStat(A, B, option)
}
\arguments{
\item{A}{is a distance matrix or a n*d data matrix; if it is not a square matrix with zeros on diagonal, it is treated as n*d data;}

\item{B}{is a second distance matrix or a n*d data matrix, with the same distance matrix check as A;}

\item{option}{is a string that specifies which global correlation to build up-on, including 'mgc','dcor','mantel', and 'rank'.}
}
\value{
A list contains the following output:

statMGC is the sample MGC statistic within [-1,1];

localCorr consists of all local correlations by double matrix index;

optimalScale the estimated optimal scale in matrix single index.
}
\description{
The main function that computes the MGC measure between two datasets:
It first computes all local correlations,
then thresholding and use the smoothed maximal of the local correlations.
}
