% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkTest.R
\name{NetworkTest.diffusion.stat}
\alias{NetworkTest.diffusion.stat}
\title{This function prints out the distance-based independence test (MGC, mCorr, HHG) statistics.
The statistics depend on the choice of diffusion map embeddings.}
\usage{
NetworkTest.diffusion.stat(G, X, option, diffusion, t.range, n.perm)
}
\arguments{
\item{G}{is a igraph object having n nodes;}

\item{X}{is a vector of length [n] or [nxn] matrix;}

\item{option}{specifies the types of distance-based independence test statistic.
\describe{
   \item{'1'}{use the MGC.}
   \item{'2'}{use the mCorr.}
   \item{'3'}{use the HHG.}
}}

\item{diffusion}{is interpreted as:
\describe{
   \item{'TRUE'}{use the diffusion map embeddings.}
   \item{'FALSE'}{use the columns of adjacency matrix as the ad-hoc graph embeddings.}
}}

\item{t.range}{is a range of Markov iteration times applied for diffusion map embeddings;}

\item{n.perm}{is the number of permutation samples.}
}
\value{
p-value of the test if \code{diffusion} is \code{"FALSE"};

a collection of statistics under the alternative and \code{n.perm} number of nulls across \code{t.range} if \code{diffusion} is \code{"TRUE"}.
}
\description{
This function prints out the distance-based independence test (MGC, mCorr, HHG) statistics.
The statistics depend on the choice of diffusion map embeddings.
}
\author{
Youjin Lee
}
