% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkTest.R
\name{NetworkTest}
\alias{NetworkTest}
\title{Multiscale Network Test}
\usage{
NetworkTest(G, X, option = "mgc", t.max = 5, n.perm = 100,
  default.q = 10, default.t = 3)
}
\arguments{
\item{G}{the input igraph object.}

\item{X}{the input nodal attributes.}

\item{option}{is a string that specifies which global correlation to build up-on. Defaults to \code{'mgc'}.
\describe{
   \item{\code{'mgc'}}{uses the MGC global correlation.}
   \item{\code{'dcor'}}{uses the dcor global correlation.}
   \item{\code{'hhg'}}{uses Heller-Heller-Gorfine (HHG) tests.}
}}

\item{t.max}{the maximum diffusion time. Diffusion time starts from 0,1,.. \code{t.max}. Defaults to \code{t.max} = 5.}

\item{n.perm}{the number of permutation samples.  Defaults to \code{n.perm} = 100.}

\item{default.q}{dimension of graph embeddings. Defaults to \code{default.q} = 10.}

\item{default.t}{default diffusion time at optimal when no significant signal was found. Defaults to \code{default.t} = 3.}
}
\value{
A list containing the following:
\item{\code{pval}}{is p-value under diffusion map at \code{optimal.t}.}
\item{\code{optimal.t}}{diffusion time at smoothed maximum among multiscale statistics.}
\item{\code{dimension}}{a dimension of diffusion map embedding.}
\item{\code{localmap}}{a list of local correlation map (in case of \code{option} = '\code{mgc}').}
\item{\code{localscale}}{a list of optimal local scale (row, column) in each local correlation map (in case of \code{option} = '\code{mgc}').}
}
\description{
This function provides multiscale test statistics and p-values, and those under optimal signal
in testing independence between edge connectivity in graph \code{G} and nodal attributes \code{X} using diffusion map embeddings. Default test method is \code{mgc}.
}
\examples{
library(mgc)
library(igraphdata)
data(karate)
G <- karate
X <- V(karate)$Faction
result <- NetworkTest(G, X, option = 'mgc', t.max = 5)
print(result$pval) # p-value
print(result$optimal.t) # optimal diffusion time

}
\author{
Youjin Lee
}
