% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkTest.R
\name{print.lambda}
\alias{print.lambda}
\title{Eigenvalues of diffusion map}
\usage{
print.lambda(L, times = 1)
}
\arguments{
\item{L}{[n x n] Laplacian matrix.}

\item{times}{a vector of diffusion times. Defaults to 1.}
}
\value{
a list of eigenvalues at each diffusion time.
}
\description{
This auxiliary function provides an eigenvalues of the diffusion maps at given diffusion time.
}
\author{
Youjin Lee
}
