% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalT.R
\name{print.stat.optimal}
\alias{print.stat.optimal}
\title{This function empirically searches for data-adaptive optimal scale t for diffusion map embeddings.
Input data is a list of statistics under the alternative and those under the null (from permutation).}
\usage{
\method{print}{stat.optimal}(stat.list, default.t)
}
\arguments{
\item{stat.list}{is interpreted as:
\describe{
   \item{list of length [2]}{List of a length-[q] vector and [nr x q] matrix comprised of statistics under the alternative and under the null.}
}}

\item{default}{is the position of default Markov time [t] among [q].}
}
\value{
p-value using the opitmal scale; a collection of statistics; optiaml t chosen for the alternative and for [nr] cases of null.
}
\description{
This function empirically searches for data-adaptive optimal scale t for diffusion map embeddings.
Input data is a list of statistics under the alternative and those under the null (from permutation).
}
\author{
Youjin Lee
}
