% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_isolates.R
\name{remove.isolates}
\alias{remove.isolates}
\title{Remove Isolates}
\usage{
remove.isolates(X, Y, is.dist = FALSE)
}
\arguments{
\item{X}{is interpreted as:
\describe{
   \item{a \code{[n x n]} distance matrix}{X is a square matrix with zeros on diagonal for \code{n} samples.}
   \item{a \code{[n x d]} data matrix}{X is a data matrix with \code{n} samples in \code{d} dimensions.}
}}

\item{Y}{\code{[n]} a vector containing the sample ids for our \code{n} samples.}

\item{is.dist}{a boolean indicating whether your \code{X} input is a distance matrix or not. Defaults to \code{FALSE}.}
}
\description{
A function to remove isolates from a dataset, given a data matrix or a distance matrix.
}
\author{
Eric Bridgeford
}
\keyword{internal}
