% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCSampleStat.R
\name{Thresholding}
\alias{Thresholding}
\title{An auxiliary function that finds a region of significance in the local correlation map by thresholding.}
\usage{
Thresholding(localCorr, m, n, sz)
}
\arguments{
\item{localCorr}{is all local correlations}

\item{m}{is the number of rows of localCorr}

\item{n}{is the number of columns of localCorr}

\item{sz}{is the sample size of original data (which may not equal m or n in case of repeating data).}
}
\value{
R is a binary matrix of size m and n, with 1's indicating the significant region.
}
\description{
An auxiliary function that finds a region of significance in the local correlation map by thresholding.
}
\author{
Eric Bridgeford and C. Shen
}
\keyword{internal}
