% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDCVBlockBootMulti.R
\name{cdcv.multi.test}
\alias{cdcv.multi.test}
\title{Full test of independence of multivariate time series X and Y.
Uses cross-distance covariance and circular or stationary bootstrap.
For univariate data, use cdcv.univ.test.}
\usage{
cdcv.multi.test(X, Y, M = NA, num_boot = 100,
  boot_type = "stationary", block_size = NA)
}
\arguments{
\item{X}{Time series as n x d_X matrix.}

\item{Y}{Time series as n x d_Y matrix.}

\item{M}{Maximum lag to consider for cross-distance covariance. Defaults to \code{log(n)}.}

\item{num_boot}{Number of bootstrapped samples. Defaults to \code{100}.}

\item{boot_type}{"stationary" or "circular" bootstrap. Defaults to \code{stationary}.}

\item{block_size}{Block size or mean block size. Defaults to \code{sqrt(n)}.}
}
\value{
A list containing the following:
\item{\code{pCDCV}}{P-value of CDCV}
\item{\code{statCDCV}}{test statistic value}
\item{\code{unbiased}}{Whether to use the unbiased or biased estimate of CDCV}
\item{\code{M}}{the max lead/lag considered}
}
\description{
Full test of independence of multivariate time series X and Y.
Uses cross-distance covariance and circular or stationary bootstrap.
For univariate data, use cdcv.univ.test.
}
