% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrStat.R
\name{discr.discr}
\alias{discr.discr}
\title{Discriminability}
\usage{
discr.discr(D, ids, thresh = 0, verbose = FALSE)
}
\arguments{
\item{D}{is interpreted as:
\describe{
   \item{a [n x n] distance matrix}{D is a square matrix with zeros on diagonal}
   \item{a [n x d] data matrix}{Otherwise}
}}

\item{ids}{[n]: a vector containing the labels for our n observations.}

\item{remove_outliers=TRUE}{boolean indicating whether to ignore observations with rdf below a certain cutoff.}

\item{thresh=0}{[1]: the threshold below which to ignore observations. If thresh > 0, ignores observations where the rdf is < thresh in the discriminability computation.}

\item{verbose=FALSE}{a boolean indicating whether to:
\describe{
   \item{TRUE}{print output to console}
   \item{FALSE}{Do not print output to console}
}}
}
\value{
discr the discriminability statistic.
}
\description{
A function for computing the discriminability from a distance matrix and a set of associated labels.
}
\seealso{
{discr.distance}
}
\author{
Eric Bridgeford and Gregory Kiar
}
