% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrDistance.R
\name{discr.distance}
\alias{discr.distance}
\title{Distance}
\usage{
discr.distance(X, method = "euclidean")
}
\arguments{
\item{X}{\code{[n x d]} a data matrix for \code{d} samples of \code{d} variables.}

\item{method}{the method for computing distances. Defaults to \code{'euclidean'}. See \link[stats]{dist} for details.}
}
\value{
a \code{[n x n]} distance matrix.
\code{[n, n]} distance matrix indicating the pairwise distances between all samples passed in.
}
\description{
A function that returns a distance matrix given a collection of observations.
}
\author{
Eric Bridgeford
}
