% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrDistance.R
\name{discr.distance}
\alias{discr.distance}
\title{Distance}
\usage{
discr.distance(X, method = "2")
}
\arguments{
\item{X}{\code{[n, d]} a data matrix for n samples of d variables.}

\item{method}{the method for computing distances. Defaults to \code{'2'}.
\describe{
   \item{'2'}{use the 2 norm between pairs of observations}
}}
}
\value{
a \code{[n, n]} distance matrix indicating the pairwise distances between all samples passed in.
}
\description{
A function that returns a distance matrix given a collection of observations.
}
\author{
Eric Bridgeford
}
