% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrStat.R
\name{discr.mnr}
\alias{discr.mnr}
\title{Discriminability Mean Normalized Rank}
\usage{
discr.mnr(rdf)
}
\arguments{
\item{rdf}{the reliability densities.}
}
\value{
the mnr.
}
\description{
Discriminability Mean Normalized Rank
}
\keyword{internal}
