% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrStat.R
\name{discr.rdf}
\alias{discr.rdf}
\title{Reliability Density Function}
\usage{
discr.rdf(X, ids)
}
\arguments{
\item{X}{\code{[n, n]} a distance matrix for n samples.}

\item{ids}{\code{[n]} a vector containing the label ids for each sample.}
}
\value{
\code{[n]} vector of the reliability per sample.
}
\description{
A function for computing the reliability density function of a dataset.
}
\author{
Eric Bridgeford
}
\keyword{internal}
