% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrStat.R
\name{discr.validator}
\alias{discr.validator}
\title{Utility Validator}
\usage{
discr.validator(X, Y, is.dist)
}
\arguments{
\item{X}{is interpreted as:
\describe{
   \item{a \code{[n x d]} data matrix}{X is a data matrix with \code{n} samples in \code{d} dimensions, if flag \code{is.dist=FALSE}.}
   \item{a \code{[n x n]} distance matrix}{X is a distance matrix. Use flag \code{is.dist=TRUE}.}
}}

\item{Y}{\code{[n]} a vector containing the sample ids for our \code{n} samples.}

\item{is.dist}{a boolean indicating whether your \code{X} input is a distance matrix or not.}
}
\description{
A script that validates that data inputs are correct.
}
