% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmaps.R
\name{dmap.q}
\alias{dmap.q}
\title{This function prints out the diffusion maps having a fixed dimension from a normalized graph Laplacian.}
\usage{
dmap.q(L, times, q)
}
\arguments{
\item{L}{is a normalized graph Laplacian having n nodes;}

\item{times}{is is a range of Markov iteration times applied for diffusion map embeddings;}
}
\value{
a list of diffusion maps.
}
\description{
This function prints out the diffusion maps having a fixed dimension from a normalized graph Laplacian.
}
\author{
Youjin Lee
}
