% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCKSample.R
\name{mgc.ksample}
\alias{mgc.ksample}
\title{MGC K Sample Testing}
\usage{
mgc.ksample(X, Y, mgc.opts = list(), ...)
}
\arguments{
\item{X}{\code{[n, d]} the data with \code{n} samples in \code{d} dimensions.}

\item{Y}{\code{[n]} the labels of the samples with \code{K} unique labels.}

\item{mgc.opts}{Arguments to pass to MGC. See \code{\link{MGC}} for details.}

\item{...}{trailing args.}
}
\value{
A list containing the following:
\item{\code{pMGC}}{P-value of MGC}
\item{\code{statMGC}}{is the sample MGC statistic within \code{[-1,1]}}
\item{\code{pLocalCorr}}{P-value of the local correlations by double matrix index}
\item{\code{localCorr}}{the local correlations}
\item{\code{optimalScale}}{the optimal scale identified by MGC}
}
\description{
MGC K Sample Testing provides a wrapper for MGC Sample testing under the constraint that the Ys here are
categorical labels with K possible sample ids. This function uses a 0-1 loss for the Ys. To use a custom distance
function, use \code{mgc.test} with your custom distance function as \code{Y}.
}
\author{
Eric Bridgeford
}
