% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCLocalCorr.R
\name{mgc.localcorr}
\alias{mgc.localcorr}
\title{Compute all local correlation coefficients in O(n^2 log n)}
\usage{
mgc.localcorr(X, Y, option = "mgc")
}
\arguments{
\item{X}{is a [n, n] distance matrix or a [n, d] data matrix; if it is not a square matrix with zeros on diagonal, it is treated as n*d data;}

\item{Y}{is a second [nxn] distance matrix or a [nxd] data matrix; if it is not a square matrix with zeros on diagonal, it is treated as [nxd] data}

\item{option='mgc'}{is a string that specifies which global correlation to build up-on.
\describe{
   \item{'mgc'}{use the MGC global correlation.}
   \item{'dcor'}{use the dcor global correlation.}
   \item{'mantel'}{use the mantel global correlation.}
   \item{'rank'}{use the rank global correlation.}
}}
}
\value{
A list contains the following output:

corr consists of all local correlations within [-1,1] by double matrix index;

varX contains all local variances for X; varY contains all local covariances for Y.
}
\description{
Compute all local correlation coefficients in O(n^2 log n)
}
\author{
C. Shen
}
