% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{mgc.plot.plot_matrix}
\alias{mgc.plot.plot_matrix}
\title{Matrix plot}
\usage{
mgc.plot.plot_matrix(mtx, title = "", xlabel = "", ylabel = "",
  legend.name = "metric", legend.show = TRUE, font.size = 12,
  limits = NaN, vfactor = FALSE, ffactor = FALSE)
}
\arguments{
\item{mtx}{[n, n] or [n, d]: the input. Can either be a square matrix or a data matrix.}

\item{title="":}{the title for the square plot.}

\item{xlabel="":}{the x label for the square plot.}

\item{ylabel="":}{the y label for the square plot.}

\item{legend.name="":}{the legend title for the square plot.}

\item{legend.show=TRUE:}{whether to show the legend on the plot.}

\item{font.size=12:}{the default font size for the plot text. Axis/legend text is font.size - 2.}

\item{limits=NaN}{if the limits are not specified, limits are c(min(mtx), max(mtx)) for color fill.}

\item{vfactor=FALSE}{if TRUE, interpret the vertex labels as a factor.}

\item{ffactor=FALSE}{if TRUE, interpret the fill colors as a factor.}
}
\description{
A function that plots a matrix.
}
\author{
Eric Bridgeford
}
