% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCSampleStat.R
\name{mgc.sample}
\alias{mgc.sample}
\title{The main function that computes the MGC measure between two datasets:
It first computes all local correlations,
then use the maximal statistic among all local correlations based on thresholding.}
\usage{
mgc.sample(A, B, option = "mgc")
}
\arguments{
\item{A}{[n, d] is a distance matrix or a n*d data matrix; if it is not a square matrix with zeros on diagonal, it is treated as n*d data;}

\item{B}{[n, d] is a second distance matrix or a n*d data matrix, with the same distance matrix check as A;}

\item{option='mgc'}{is a string that specifies which global correlation to build up-on.
\describe{
   \item{'mgc'}{use the MGC global correlation.}
   \item{'dcor'}{use the dcor global correlation.}
   \item{'mantel'}{use the mantel global correlation.}
   \item{'rank'}{use the rank global correlation.}
}}
}
\value{
statMGC is the sample MGC statistic within [-1,1];

localCorr consists of all local correlations by double matrix index;

optimalScale the estimated optimal scale in matrix single index.
}
\description{
The main function that computes the MGC measure between two datasets:
It first computes all local correlations,
then use the maximal statistic among all local correlations based on thresholding.
}
\author{
C. Shen
}
