% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCSampleStat.R
\name{mgc.sample}
\alias{mgc.sample}
\title{The main function that computes the MGC measure between two datasets:
It first computes all local correlations,
then use the maximal statistic among all local correlations based on thresholding.}
\usage{
mgc.sample(A, B, option = "mgc")
}
\arguments{
\item{A}{is interpreted as:
\describe{
   \item{a [nxn] distance matrix}{A is a square matrix with zeros on diagonal}
   \item{a [nxd] data matrix}{Otherwise}
}}

\item{B}{is interpreted as:
\describe{
   \item{a [nxn] distance matrix}{B is a square matrix with zeros on diagonal}
   \item{a [nxd] data matrix}{Otherwise}
}}

\item{option='mgc'}{is a string that specifies which global correlation to build up-on.
\describe{
   \item{'mgc'}{use the MGC global correlation.}
   \item{'dcor'}{use the dcor global correlation.}
   \item{'mantel'}{use the mantel global correlation.}
   \item{'rank'}{use the rank global correlation.}
}}
}
\value{
statMGC is the sample MGC statistic within [-1,1];

localCorr consists of all local correlations by double matrix index;

optimalScale the estimated optimal scale in matrix single index.
}
\description{
The main function that computes the MGC measure between two datasets:
It first computes all local correlations,
then use the maximal statistic among all local correlations based on thresholding.
}
\author{
C. Shen
}
