% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGCPermutationTest.R
\name{mgc.test}
\alias{mgc.test}
\title{MGC Permutation Test}
\usage{
mgc.test(X, Y, rep = 1000, option = "mgc")
}
\arguments{
\item{X}{is interpreted as:
\describe{
   \item{a [n x n] distance matrix}{X is a square matrix with zeros on diagonal}
   \item{a [n x d] data matrix}{Otherwise}
}}

\item{Y}{is interpreted as:
\describe{
   \item{a [n x n] distance matrix}{Y is a square matrix with zeros on diagonal}
   \item{a [n x d] data matrix}{Otherwise}
}}

\item{rep=1000}{specifies the number of replicates to use for the permutation test;}

\item{option='mgc'}{is a string that specifies which global correlation to build up-on.
\describe{
   \item{'mgc'}{use the MGC global correlation.}
   \item{'dcor'}{use the dcor global correlation.}
   \item{'mantel'}{use the mantel global correlation.}
   \item{'rank'}{use the rank global correlation.}
}}
}
\value{
P-value and test statistic of MGC;

All local p-values by double matrix index, all local correlations by double matrix index, and the estimated optimal scales as matrix single indices.

Note that one should avoid report positive discovery via minimizing individual p-values of local correlations,
unless corrected for multiple testing problem.
}
\description{
The main function that tests independent between two data sets by MGC and permutation test.
}
\author{
C. Shen
}
