% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmaps.R
\name{print.lambda}
\alias{print.lambda}
\title{This function prints out eigenvalues of a normalized graph Laplacian at given times.
A set of returned eigenvalues are often used for finding the dimension of embeddings.}
\usage{
\method{print}{lambda}(L, times)
}
\arguments{
\item{L}{is a normalized graph Laplacian having n nodes;}

\item{times}{is is a range of Markov iteration times applied for diffusion map embeddings;}
}
\value{
a set of eigenvalues.
}
\description{
This function prints out eigenvalues of a normalized graph Laplacian at given times.
A set of returned eigenvalues are often used for finding the dimension of embeddings.
}
\author{
Youjin Lee
}
