% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkTest.R
\name{print.stat.optimal}
\alias{print.stat.optimal}
\title{Smoothed maximum statistics}
\usage{
print.stat.optimal(stat.list, default.t = 3)
}
\arguments{
\item{stat.list}{a list containing following statistics:
\describe{
\item{\code{stats}}{\code{[t]} statistics across diffusion times.}
\item{\code{stats.null}}{\code{[n.perm, t]} statistics under the permutation sample.}
}}

\item{default.t}{a default diffusion time when no significant signal was found. Defaults to \code{default.t} = 3.}
}
\value{
A list containing the following:
\describe{
\item{\code{pval}}{is p-value under diffusion map at \code{alt.t}.}
\item{\code{alt.stat}}{a test statistic at \code{alt.t}.}
\item{\code{alt.t}}{diffusion time at smoothed maximum among multiscale statistics.}
\item{\code{null.stat}}{test statistics at \code{null.t} upon permutation sample.}
\item{\code{null.t}}{diffusion time at smoothed maximum among multiscale statistics under the null.}
}
}
\description{
This function prints out the statistic at smoothed maximum diffusion time,
 its p-value and the statistics under the null based on the permutation.
}
\author{
Youjin Lee
}
